fix benchmarks for CI, when run without all features
