+++ relay-cardinality/Cargo.toml
@@ -8,6 +8,7 @@ version = "23.10.1"
 edition = "2021"
 license-file = "../LICENSE"
 publish = false
+autobenches = false
 
 [features]
 default = []
